requirements.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
blis==0.4.1
boto3==1.9.176
botocore==1.12.176
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorgram.py==1.2.0
colorharmonies==1.0.5
croniter==0.3.30
cycler==0.10.0
cymem==2.0.2
decorator==4.4.0
defusedxml==0.6.0
docutils==0.14
en-core-web-md==2.2.0
entrypoints==0.3
Flask==1.1.1
Flask-WTF==0.14.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
importlib-metadata==0.23
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.14.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
murmurhash==1.0.2
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==5.7.8
numpy==1.15.0
opencv-contrib-python==4.1.1.26
opencv-python==4.1.1.26
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pipenv==2018.11.26
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.2
python-crontab==2.3.8
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.0.1
qtconsole==4.5.1
requests==2.21.0
s3transfer==0.2.1
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
spacy==2.2.2
srsly==0.2.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
thinc==7.3.1
torch==1.2.0
tornado==6.0.3
tqdm==4.38.0
traitlets==4.3.2
urllib3==1.24.3
virtualenv==16.7.1
virtualenv-clone==0.5.3
wasabi==0.4.0
wcwidth==0.1.7
webcolors==1.10
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.4.2
wrapt==1.11.2
WTForms==2.2.1
zipp==0.6.0
